refactor: implementing tag variants across entire application
