+++ components/InfoCard/index.tsx
@@ -1,18 +1,17 @@
 // COMPONENTS
-import Tag from "../CourseCard/Tag"
+import { Tag as TagType } from "@/.contentlayer/generated";
 // UTILS
 import { cn } from "@/utils/helpers";
+import { tagVariants, Tag } from "../ui/Tag";
 
 type InfoCardProps = {
-  icon?: boolean;
-  hastags?: string[];
+  hastags?: TagType[];
   title: string;
   description: string;
   className?: string;
 }
 
 export default async function InfoCard({ 
-  icon, 
   hastags, 
   title, 
   description, 
@@ -26,11 +25,11 @@ export default async function InfoCard({
     >
       {hastags && hastags.length && (
         <div className="flex justify-start items-center gap-2">
-          {hastags && hastags.map((hastag) => (
+          {hastags && hastags.map(({ title, variant}) => (
             <Tag 
-              key={hastag} 
-              tag={hastag} 
-              className="bg-primary-black"
+              key={title} 
+              tag={title} 
+              className={cn(tagVariants({ variant }))}
             /> 
           ))}
         </div>
