+++ content/courses/nextjs/index.mdx
@@ -3,12 +3,17 @@ title: React Fundamentals Title Go
 description: All the essentials in a tight packed project. Checking how many chars
 weight: 1
 date: "2023-01-08"
-tags: 
-    - pro
-
+tags:
+  - 
+    title: Free
+    variant: primary
 stack: 
-    - FullStack
-    - NextJS
+    - 
+        title: NextJS
+        variant: alternateTwo
+    - 
+        title: NextJS
+        variant: alternateTwoOutlined
 ---
 
 This is next.js first course
