+++ content/courses/svelt/index.mdx
@@ -3,6 +3,10 @@ title: Welcome to Svelt course
 description: Overview of next.js latest features.
 weight: 1
 date: "2023-01-09"
+stack: 
+    - 
+        title: FullStack
+        variant: alternateTwo
 ---
 
 This is next.js first course
