+++ contentlayer.config.ts
@@ -53,8 +53,8 @@ export const Course = defineDocumentType(() => ({
       youtube: { type: 'string' },
       video_length: { type: 'string' },
       date: { type: 'date', required: true },
-      tags: { type: 'list', of: { type: 'string' } },
-      stack: { type: 'list', of: { type: 'string' } },
+      tags: { type: 'list', of: Tag },
+      stack: { type: 'list', of: Tag },
       lastmod: { type: 'date' },
       free: { type: 'boolean' },
     },
